third_party/WebKit/LayoutTests/fast/dom/shadow/css-cue-for-video-in-shadow-expected.txt

Test that the cue is not styled when video is in a shadow tree and style is in a document.
EVENT(canplaythrough)
EVENT(seeked)
EXPECTED (getComputedStyle(cueNode).color != 'rgb(255, 0, 0)') OK
EXPECTED (getComputedStyle(cueNode).color != 'rgb(0, 128, 0)') OK
EXPECTED (getComputedStyle(cueNode).color != 'rgb(255, 0, 0)') OK
END OF TEST
